requirements-dev.txt
pytest
pytest-cov
hypothesis
ruff
bandit
black
pyyaml
jsonschema
